pyproject.toml
[project]
name = "shaderbox"
version = "0.1.0"
authors = [
    { name = "Alexey Karnachev", email = "alekseykarnachev@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "glfw>=2.8.0",
    "moderngl>=5.12.0",
    "numpy>=2.2.4",
    "pillow>=11.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["shaderbox"]

[tool.uv]
dev-dependencies = [ ]

